don't open hostfile in binary mode (SOFTWARE-5197)
This "open(args.hostfile, 'rb')" has been there since the initial
import commit; so it seems to be a holdover from the python2 days.

The result is the lines read from the file are stored as bytes
(which used to be the same as str in python2), instead of unicode
(which of course is the same as str in python3).

The result is the common_name/hostname passed into the cert_utils.Csr()
__init__ is a bytes, but it is used to construct a path with other
regular str (ie unicode) strings, which raises a TypeError for
trying to concat str to bytes.

(That's what it looks like is happening to me, anyway.)